+++ plugins/hydra_nevergrad_sweeper/MANIFEST.in
