+++ plugins/hydra_nevergrad_sweeper/example/config.yaml
@@ -1,5 +1,5 @@
 defaults:
-  - hydra/sweeper: nevergrad-minimizer
+  - hydra/sweeper: nevergrad-sweeper
 hydra:
   sweeper:
     params:
