+++ plugins/hydra_nevergrad_sweeper/hydra_plugins/hydra_nevergrad_sweeper/__init__.py
@@ -1,3 +1,3 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
-from .core import NevergradMinimizer
-__all__ = ["NevergradMinimizer"]
+from .core import NevergradSweeper
+__all__ = ["NevergradSweeper"]
